backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
pydantic
python-multipart
